Overview && Image Gallery new functionality -> Render 5 plus scroll
Implemented ability to have max of 5 images rendered dynamically in gallery
CSS grid section, added function to detect whether or not the reference the gallery
was pointing to was big enough for a need to add a button for scrolling. If the reference is greater than 5, a button for scrolling will appear. Added start and end limits for the gallery to detect dynamically which chunk of the reference to render and a way to detect if at the end of the reference length, go back to the start.